i got the save to work but the modals friggin pmo im just an idiotidkt whythis is taking me so logna aweaweawejopawejopaweawejopaejoaweawe
